refactor diffusion
